+++ EntityTest/CFrustum.hpp
@@ -2,6 +2,8 @@
 
 #include <array>
 
+#include "Types.hpp"
+
 #include "CPlane.hpp"
 
 class CFrustum
@@ -9,7 +11,7 @@ class CFrustum
 public:
 	explicit CFrustum( const glm::mat4 &viewProjectionMatrix );
 
-	bool IsSphereInside( const glm::vec3 &position, const float boundingSphereRadius ) const;
+	bool IsSphereInside( const glm::vec3 &position, const f16 boundingSphereRadius ) const;
 
 	const std::array< CPlane, 6 > &Planes() const;
